+++ protocols/gossipsub/CHANGELOG.md
@@ -15,8 +15,7 @@
 - Fix underflow when shuffling peers after prunning.
   See [PR 6183](https://github.com/libp2p/rust-libp2p/pull/6183)
 
-
-- Implement gossipsub 1.3 partial messages extension
+- Implement gossipsub 1.3 partial messages extension.
   See [PR XXXX](https://github.com/libp2p/rust-libp2p/pull/XXXX)
 
 - Remove peer penalty for duplicate messages.